requirements.txt
pandas==1.4.1
scikit-learn==1.0.2
nltk==3.7
stopwordsiso==0.6.1
django==4.0.2
djangorestframework
markdown
django-filter
django-cors-headers
chartdet